WIP: split SideBar into separate view and business logic components
