+++ packages/web-app-files/tests/unit/components/SideBar/SideBar.spec.js
@@ -8,6 +8,7 @@ import Files from '@/__fixtures__/files'
 import merge from 'lodash-es/merge'
 import { buildResource, renameResource } from '@files/src/helpers/resources'
 
+import InnerSideBar from 'web-pkg/src/components/sidebar/SideBar.vue'
 import SideBar from '@files/src/components/SideBar/SideBar.vue'
 import { createLocationSpaces } from '../../../../src/router'
 
@@ -226,7 +227,10 @@ function createWrapper({
       }
     }),
     localVue,
-    stubs,
+    stubs: {
+      ...stubs,
+      SideBar: InnerSideBar
+    },
     directives: {
       'click-outside': jest.fn()
     },
